% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_approx.R
\docType{data}
\name{local_approx}
\alias{local_approx}
\title{Square Each Element of a Numeric Vector (Rcpp Version)}
\format{
An object of class \code{function} of length 1.
}
\usage{
local_approx(data, start_p, m, neighbors)
}
\arguments{
\item{data}{A numeric vector.}

\item{start_p}{Starting point of the data (after this point predictions start).}

\item{m}{Embedding dimension.}

\item{neighbors}{Number of neighbors to consider.}
}
\value{
A forecast of the numeric vector.
}
\description{
This function squares each element of a numeric vector using Rcpp.
}
\examples{
#'
\dontrun{
  local_approx(data = 1:10, start_p = 2, m = 2, neighbors = 1)
}
}
\keyword{datasets}
