% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_space.R
\docType{package}
\name{phase_space}
\alias{phase_space}
\title{Natural Visibility Graph}
\format{
An object of class \code{function} of length 1.
}
\usage{
phase_space(x,m,tau)
}
\description{
This procedure uses time series data and convert it
into vector/multi-variate time series data
}
\note{
This procedure returns diagonal elements of adjacency matrix as connected (1).
}
\examples{
phase_space(1:10,3,1)


}
\keyword{'Chaos'}
\keyword{Phase}
\keyword{Space,}
