% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VGA.R
\docType{package}
\name{VGA}
\alias{VGA}
\title{Natural Visibility Graph}
\format{
An object of class \code{function} of length 1.
}
\usage{
VGA(t)
}
\description{
This procedure uses time series vector and convert it
into network called temporal network
}
\note{
This procedure returns diagonal elements of adjacency matrix as connected (1).
}
\examples{
VGA(sample(10))


}
\keyword{'Complex}
\keyword{'Visibility}
\keyword{Graphs',}
\keyword{Graphs,}
\keyword{NVG,}
\keyword{Network'}
