% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaosdata.R
\docType{data}
\name{chaosdata}
\alias{chaosdata}
\title{Chaos Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10000 rows and 3 columns.
}
\usage{
data(chaosdata)
}
\description{
Data from three different chaotic systems are loaded.
Each has ten thousand data points. The data imports
Rossler system, Lorenz Data, and Mackey and Glass data into environment
}
\examples{
data(chaosdata)

}
\keyword{datasets}
