% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnn_algo.R
\docType{package}
\name{fnn_algo}
\alias{fnn_algo}
\title{False Nearset Neighbour}
\format{
An object of class \code{function} of length 1.
}
\usage{
fnn_algo(x,tau=1,mmax=5)
}
\description{
False Nearest Neighbour a dimensionality estimation method.
This method uses delay value to reconstruct phase space and identifies dimension
based on true neighbours
Rossler system, Lorenz Data, and Mackey and Glass data into environment
}
\examples{
data(chaosdata)

fnn_algo(chaosdata$Lorenz,tau=3,mmax=10)

}
\keyword{Dimension,}
\keyword{FNN,}
\keyword{Phase-Space}
