% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VGA}
\alias{VGA}
\title{Converts Time series to Complex Networks}
\usage{
VGA(t)
}
\arguments{
\item{t}{Time series vector.}
}
\value{
Returns adjacency matrix of the time series.
}
\description{
This function converts time domain to network domain
}
\note{
This procedure returns diagonal elements of adjacency matrix as connected (1).
}
\examples{
VGA(sample(10))
}
