% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vvga}
\alias{vvga}
\title{Converts multivariate Time series data in to Complex Networks}
\usage{
vvga(v)
}
\arguments{
\item{v}{multivariate time series vector.}
}
\value{
Returns adjacency matrix of the time series.
}
\description{
This function takes multiple variables and convert into complex network
}
\note{
This procedure returns diagonal elements of adjacency matrix as connected (1).

When the evaluation function is equal, a connection is establishes in the matrix (not in accordance with Lacasa et al., 2009)
}
\examples{
VGA(sample(10))
}
\references{
Lacasa., Sivakumar, B (2009). XXX JOH, V(3),1.
}
